fix tux jumping always full height
SVN-Revision: 2138

+++ src/gameobjs.cpp
@@ -452,7 +452,9 @@ SmokeCloud::draw(DrawingContext& context)
   img_smoke_cloud->draw(context, position, LAYER_OBJECTS+1);
 }
 
-Particles::Particles(const Vector& epicenter, int min_angle, int max_angle, const Vector& initial_velocity, const Vector& acceleration, int number, Color color_, int size_, int life_time, int drawing_layer_)
+Particles::Particles(const Vector& epicenter, int min_angle, int max_angle,
+        const Vector& initial_velocity, const Vector& acceleration, int number,
+        Color color_, int size_, float life_time, int drawing_layer_)
   : accel(acceleration), color(color_), size(size_), drawing_layer(drawing_layer_)
 {
   if(life_time == 0) {
